Use legacy syntax for string "split" method
In Python 3, we may split a string s into at most 3 fields, separated by
whitespaces, with

    s.split(maxsplit=2)

Unfortunately, in Python 2, the split method does not take keyword
arguments, and the above has to be expressed as

    s.split(None, 2)

Since we want to be compatible with both Python 2 and Python 3, this
patch replaces the Python 3 call with the legacy version. Since this
syntax is much less readable, the logic is explained in a comment.