Clarify maxsplit parameter in str split method
Added a comment to clarify a question to came up during code review.
When giving the `maxsplit` parameter in the `split` method of a string,
it specifies not the maximum length of the resulting list, but the
number of splits that are performed. Doing n splits results in n+1
items.

>>> s = "1 2 3 4"
>>> s.split()
['1', '2', '3', '4']
>>> s.split(maxsplit=2)
['1', '2', '3 4']