Fix bug where backlashes were dropped from options
While backslashes may be used to escape quotes in options, they should
be kept as backslashes if they're not followed by a quote. The following
is a perfectly valid line in an authorized_keys file:

    command="echo \"\HELLO\"" ssh-rsa AAAAB...

in order to print \HELLO upon connection. Currently, the backslash
before HELLO would have been dropped by the parser.